+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 Release notes
-==============
+============== 
+
+### Unreleased
+
+* The invalid value is now included in the error mesage for attribute parse errors. The value is limited to 200 characters. * Parse error messages for attributes defined with overloads that return user-defined error messages (e.g. `Result<_, string>`) are generally of the form `Attribute 'name' got invalid value 'invalidValue': <User-defined error message>`
 
 ### 0.21.2 (2022-10-28)